refactor: consolidate integration test code
