+++ crates/deltalake-test/Cargo.toml
@@ -10,10 +10,12 @@ chrono = { workspace = true, default-features = false, features = ["clock"] }
 deltalake-core = { path = "../deltalake-core" }
 dotenvy = "0"
 fs_extra = "1.3.0"
+object_store = { workspace = true }
 rand = "0.8"
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tempfile = "3"
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 
 [features]
 default = []
