+++ crates/deltalake-test/src/lib.rs
@@ -15,9 +15,14 @@ use std::sync::Arc;
 use tempfile::TempDir;
 
 pub mod clock;
+pub mod concurrent;
 #[cfg(feature = "datafusion")]
 pub mod datafusion;
+pub mod read;
 pub mod utils;
+
+pub use concurrent::test_concurrent_writes;
+pub use read::test_read_tables;
 
 #[derive(Default)]
 pub struct TestContext {
